feat(core): filter consolidated by resource
Ref. metriport/metriport-internal#2215
