+++ packages/core/src/command/consolidated/consolidated-filter-by-date.ts
